requirements.txt
beautifulsoup4==4.11.1
deep-translator==1.9.1
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
psycopg2==2.9.5
python-decouple==3.6
requests==2.28.1
SQLAlchemy==1.4.44
fuzzysearch==0.7.3